requirements.txt
Flask==3.0.0
pyinstaller==6.3.0
pywebview==5.1